+++ src/components/Autocomplete/index.tsx
@@ -181,7 +181,20 @@ class AutocompleteInt<T> extends React.Component<AutocompleteProps<T> & FormCont
       this._element.focus();
     }
 
-    suggestionSelected ? this.hide() : this.show();
+    const hide = () => {
+      this.hide();
+      /*
+        On IE11 the focus event from this._element.focus(); triggers after this.hide() and so the suggestion list opens again after hiding it.
+        If we call this.hide() again with a timeout, it works also for IE11.
+      */
+      setTimeout(() => {
+        if (this.state.open) {
+          this.hide();
+        }
+      });
+    };
+
+    suggestionSelected ? hide() : this.show();
 
     if (typeof onChange === 'function') {
       onChange({